# Resurrect copy optimization