--- a/src/main/java/com/amazon/ion/impl/bin/AbstractIonWriter.java
+++ b/src/main/java/com/amazon/ion/impl/bin/AbstractIonWriter.java
@@ -29,12 +29,11 @@ import java.util.Date;
     }
 
     /** The cache for copy optimization checks--null if not copy optimized. */
-    private final _Private_SymtabExtendsCache symtabExtendsCache;
+    private final boolean isStreamCopyOptimized;
 
     /*package*/ AbstractIonWriter(final WriteValueOptimization optimization)
     {
-        this.symtabExtendsCache = optimization == WriteValueOptimization.COPY_OPTIMIZED
-            ? new _Private_SymtabExtendsCache() : null;
+        this.isStreamCopyOptimized = optimization == WriteValueOptimization.COPY_OPTIMIZED;
     }
 
     public final void writeValue(final IonValue value) throws IOException
@@ -56,11 +55,10 @@ import java.util.Date;
 
         if (isStreamCopyOptimized() && reader instanceof _Private_ByteTransferReader)
         {
-            if (_Private_Utils.isNonSymbolScalar(type)
-                || symtabExtendsCache.symtabsCompat(getSymbolTable(), reader.getSymbolTable()))
+            _Private_ByteTransferReader byteTransferReader = (_Private_ByteTransferReader) reader;
+            if (_Private_Utils.isNonSymbolScalar(type) || byteTransferReader.isSymbolTableCompatible(getSymbolTable()))
             {
-                // we have something we can pipe over
-                if (((_Private_ByteTransferReader) reader).transferCurrentValue(this)) {
+                if (byteTransferReader.transferCurrentValue(this)) {
                     return;
                 }
             }
@@ -220,7 +218,7 @@ import java.util.Date;
 
     public final boolean isStreamCopyOptimized()
     {
-        return symtabExtendsCache != null;
+        return isStreamCopyOptimized;
     }
 
     @SuppressWarnings("deprecation")